Renaming of solutions to Concurrency package.
