--- src/Utils/Thread/thread.md
@@ -1,1 +1,0 @@
-# OBSOLETE. DELETE
